[refactor] refactor tracing system & split macros (#6)
* [refactor] move user api macros to macros.hpp and update CI

* [refactor] cleanup old net address/endpoint code

* [fix] fix compile error on android

* [refactor] split runtime.cpp and task.cpp

* [refactor] refactor the tracing system

* [refactor] using function instead var in tracing

+++ include/ilias/io/mem.hpp
@@ -135,7 +135,7 @@ private:
         switch (origin) {
             case SeekOrigin::Begin: pos = offset; break;
             case SeekOrigin::Current: pos += offset; break;
-            case SeekOrigin::End: pos = static_cast<int64_t>(size) + offset; break;
+            case SeekOrigin::End: pos = size + offset; break;
         }
         if (pos < 0) {
             pos = 0;
